common: Refactor IPAddrToHWAddr
